Support signed Modbus metric reads with dataType mapping; preserve unsigned command semantics and telemetry boundaries; split mock and Modbus runtime fixtures; update Arduino stand config, firmware encoding, tests, and ignore rules

+++ edge_server/engineering_system/engineering_system.ino
@@ -165,7 +165,7 @@ void applyOutputs() {
 }
 
 void updateInputRegisters() {
-  inputRegisters[REG_INPUT_TEMPERATURE_X10] = (uint16_t)temperatureX10;
+  inputRegisters[REG_INPUT_TEMPERATURE_X10] = (uint16_t)((int16_t)temperatureX10);
   inputRegisters[REG_INPUT_HUMIDITY_X10] = (uint16_t)humidityX10;
   inputRegisters[REG_INPUT_LOCAL_BUTTON_PRESSED] = (uint16_t)localButtonPressed;
   inputRegisters[REG_INPUT_PUMP_ACTUAL_STATE] = (uint16_t)pumpActualState;
